Added Blood Effect
Added Blood Effect when player take damage. Also, there is a new Assets called JMO Assets, which is where I get the blood effect. There are still many other effect in it, so I left it in so we might be able to use other effect later without searching for effects.