New Cave
Kept making new styles of the cave system, decided to settle on this one. Did not manage to figure out all the lighting nor did I decorate the level. Attempted to do some basic decorating in the starting room so the players get a little idea of what will be in it.